Change tableBasic to EditStudentTable. Add style to MUI components via sx.Remove Return to Student page button in Student page and rename to Next IEP. Add gap to Box and Stack.
