+++ src/components/editStudentTable/EditStudentTable.tsx
@@ -4,7 +4,7 @@ import TableBody from "@mui/material/TableBody";
 import TableCell from "@mui/material/TableCell";
 import TableContainer from "@mui/material/TableContainer";
 import TableRow from "@mui/material/TableRow";
-import Paper from "@mui/material/Paper";
+import Box from "@mui/material/Box";
 
 const data = [
   { label: "First Name", value: "David" },
@@ -14,21 +14,29 @@ const data = [
   { label: "Email Address", value: "abc@gmail.com" },
 ];
 
-const BasicTable = () => {
+export default function EditStudentTable() {
   return (
-    <TableContainer component={Paper}>
+    <TableContainer>
       <Table aria-label="simple table">
         <TableBody>
           {data.map((row) => (
-            <TableRow key={row.label}>
-              <TableCell>{row.label}</TableCell>
-              <TableCell>{row.value}</TableCell>
+            <TableRow
+              key={row.label}
+              // sx={{
+              //   padding: "8px"
+              // }}
+            >
+              <TableCell padding="none" sx={{ padding: "1rem" }}>
+                {row.label}
+              </TableCell>
+              <Box sx={{ width: "2px", padding: "1rem" }}>:</Box>
+              <TableCell padding="none" sx={{ padding: "1rem" }}>
+                {row.value}
+              </TableCell>
             </TableRow>
           ))}
         </TableBody>
       </Table>
     </TableContainer>
   );
-};
-
-export default BasicTable;
+}
